Move the query parsing tests around a bit
Also ensure they all end with "Test" in their class and file name.
